fix: spinner heavy loading
